move Badge file to utils
